add forms for editing and add other changes

+++ Application/Emails/Commands/DeleteEmailCommand.cs
@@ -14,7 +14,9 @@ namespace PersEmails.Application.Emails.Commands
         public int Execute(IAppContext context)
         {
             var entity = context.Emails.Find(emailId);
-
+            if(entity == null)
+                return 0;
+            
             context.Emails.Remove(entity);
             
             return context.SaveChanges();
